Comentarios y refactoring
