+++ includes/vacunas/getinfo.h
@@ -1,8 +1,7 @@
-#ifndef GETINFO_H
-#define GETINFO_H
+#pragma once
 
 #include <QDialog>
-#include <vacuna.h>
+#include <vacunas/vacuna.h>
 
 namespace Ui {
 class getinfo;
@@ -17,11 +16,12 @@ public:
     ~getinfo();
 
 private slots:
+    /**
+     * @brief Botón que cierra la ventana y guarda los cambios realizados
+     */
     void on_commandLinkButton_clicked();
 
 private:
     Ui::getinfo *ui;
     vacuna& v;
 };
-
-#endif // GETINFO_H
